# Conflicts:
#	dataprovider/src/main/java/cm/aptoide/pt/dataprovider/ws/v3/PushNotificationsRequest.java
#	dataprovider/src/main/java/cm/aptoide/pt/dataprovider/ws/v3/V3.java
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/services/PullingContentService.java

+++ model/src/main/java/cm/aptoide/pt/model/v7/store/Store.java
@@ -26,16 +26,13 @@ import lombok.experimental.Accessors;
   @JsonProperty("links") private List<SocialChannel> socialChannels;
 
   public enum SocialChannelType {
-    FACEBOOK,
-    TWITTER,
-    YOUTUBE,
-    TWITCH
+    FACEBOOK, TWITTER, YOUTUBE, TWITCH
   }
 
   @Data public static class Stats {
     private int apps;
     private int subscribers;
-    private int downloads;
+    private long downloads;
   }
 
   @Data public static class Appearance {
